update by Maverick

+++ migrations/sqls/20220523223950-Orders-table-up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY KEY,
-    product_id bigint REFERENCES products(id),
+    product_id INTEGER REFERENCES products(id),
     quantity INTEGER DEFAULT 1,
-    user_id bigint REFERENCES users(id),
+    user_id INTEGER REFERENCES users(id),
     status VARCHAR(15)
 );
